backend/requirements.txt
flask==2.0.1
neo4j==4.4.0
kg-gen==0.1.0
python-docx==0.8.11
PyPDF2==2.0.0
numpy==1.21.0
pandas==1.3.0
python-dotenv==0.19.0 